Add a mechanism to round-trip yaml files
The standard PyYAML parser strips comments and other things when
it loads the data, which we want to preserve since. Use ruamel's
yaml comment preserving loader and dumpers to avoid this (though
they're not perfect, and some minor tweaks will be needed to reduce
the diffs before they're used the first time; hence this tool).

Change-Id: I159c4c15663fe744ceaf6aa100ea69b95ecfcb4e