Add a script to import a schedule
This is based on Sam's 2014-schedules repo plus the script I wrote
for doing this last year (srcomp.git/convert-schedule) but is
slightly more agnostic of the input values.

At the moment it doesn't do anything to help ensure that the result
is a particularly good schedule other than warning in one case where
it thinks the output is particularly bad.

Change-Id: I3f2a48af4b816f8d71f3312ba0eaa69383827b16