Have the schedule import search for the best fit
When given a schedule which has more slots than we actually have
teams, we want to minimise the number of empty corners in any given
match. Have the import perform this search since we don't know this
information about the seed schedules we have.

If we wanted to force a specific set of ids in the given schedule
this can still be done using the --ignore-ids mechanism previously
added as when the number of available ids is the same as the number
of teams the search only has one candidate to consider (and will
always choose that regardless of how good it is).

Change-Id: I923f7cae392ea0f808219aa816262abf8906e54d