Nicer fail if state doesn't load, offer help
Most likely the user has changed the list of teams and wants to
reschedule. If so, then removing the league.yaml file will probably
fix most things (re-scheduling once the matches have started isn't
a sane thing to do) and the import script creates one at the start
if it's missing anyway.

Change-Id: I5c64716f2895067d8025bb5e30b36137e5eaff0c