Support sorting teams' first matches by layout
This inroduces a new 'layout.yaml' file in the compstate which
records the (grouped) locations of the various teams, and orders
their first matches such that teams from the first area are in the
first match and so on. This will aid with safety checks and the
first round of shepherding.

Change-Id: Ife0f77a002cbe304961a1ec02f7534db17b88bfa