Link to some example schedules
Specifically these are the most likely to be used schedules for SR
at the moment, and are the ones I've used for each of the past few
years. Including these here so that I don't need to go searching
for them each year.

Eventually it might be nice to have some schedules pre-baked into
the software that could be selected with something like '--teams=56'.

Change-Id: If0a05d8ad914b248ca88edd056c05cae67f20015